Added file search and proof of concept LLM calls with gemini flash.
Added file search so that future tools can find and acess any file in the project.
The file content is displayed to the modefy popup.

Also implemented a class to call Gemini flash with a test classs, request changes to it and print the modefyed test class as a proof of concept of api calls